feat: add no expiration to token generated
